+++ src/AuthorityModel.java
@@ -1,32 +1,36 @@
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.HashMap;
 
+<<<<<<< HEAD
 import java.util.Map;
+=======
+import java.util.Iterator;
+>>>>>>> parent of 54654e0... Updating Authority
 
 /**
  * handles all calculations, database queries, and the overall work needed to be
  * done for handling this associated role Does NOT interact with the view class
  * directly, and also does NOT interact with the Controller class(The Controller
  * class interacts with this class, not the other way around.)
- * authority
  */
 public class AuthorityModel extends UserSuperClass {
 
 	public HashMap<String, UserSuperClass> users;
-	public HashMap<String, UserSuperClass> appoints;
 
 	//Total number of each one
 	private int patientNum;
 	private int doctorNum;
 	private int nurseNum;
 
+<<<<<<< HEAD
 	private int upcomeMonth;
 	private int upcomeDay;
 
 
 	private String[] allPatientsUsernames;
 
+=======
+>>>>>>> parent of 54654e0... Updating Authority
 	//constructor
 	public AuthorityModel(String username, char[] password, String name, String[] strings) {
 
@@ -34,6 +38,7 @@ public class AuthorityModel extends UserSuperClass {
 		setRole("authority");
 		
 	}
+<<<<<<< HEAD
 
 	//Sets up necessary information
 	//Checks for how many appointments fulfill the criteria
@@ -92,6 +97,42 @@ public class AuthorityModel extends UserSuperClass {
 		ArrayList<String> temp = Main.dbaseClass.getDepartmentList();
 		
 		return temp.size();
+=======
+	
+	public int getDepNum(){
+		ArrayList<String> temp = Main.dbaseClass.getDepartmentList();
+		
+		return temp.size();
+	}
+
+	public void getStats(){
+		users = Main.dbaseClass.getUsers();
+		Iterator<String> itr = users.keySet().iterator();
+
+		//Iterator<String> itr = users.keySet().iterator();
+		int patientNum = 0;
+		int doctorNum = 0;
+		int nurseNum = 0;
+
+		while(itr.hasNext()){
+
+			String a = itr.next();
+			String b = Main.dbase.get(a).getRole();
+			if(b.compareTo("patient") == 0){
+				patientNum++;
+			}
+			if(b.compareTo("doctor")==0){
+				doctorNum++;
+			}
+			if (b.compareTo("nurse")==0){
+				nurseNum++;
+			}
+
+		}
+		this.patientNum = patientNum;
+		this.doctorNum = doctorNum;
+		this.nurseNum = nurseNum;
+>>>>>>> parent of 54654e0... Updating Authority
 	}
 
 	//Getters and Setters
@@ -130,11 +171,6 @@ public class AuthorityModel extends UserSuperClass {
 		nurseNum = val;
 	}
 
-
-
-	public String[] getAllPatientsUsernames() {
-		return allPatientsUsernames;
-	}
 	
 
 }