+++ src/AuthorityView.java
@@ -3,6 +3,7 @@ import java.awt.BorderLayout;
 
 import javax.swing.BorderFactory;
 import javax.swing.JButton;
+import javax.swing.JComboBox;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
@@ -27,18 +28,14 @@ public class AuthorityView extends JFrame{
 	
 	private JButton btnReturn;
 
+	private JComboBox departmentDropDown;
 	
 	private JLabel welcomeLabel;
-
-
 	private JLabel departBaseLabel;
-		private JLabel departValLabel;
 	private JLabel patientTotalLabel;
-		private JLabel patientValueLabel;
-
 	private JLabel regDoctorTotalLabel;
-		private JLabel regDoctorValLabel;
 	private JLabel regNurseTotalLabel;
+<<<<<<< HEAD
 		private JLabel regNurseValLabel;
 		
 	private JLabel appointLabel;
@@ -50,15 +47,15 @@ public class AuthorityView extends JFrame{
 	private JLabel totalAppointNurLabel;
 	
 
+=======
+	private JLabel totalAppointDepLabel;
+	private JLabel totalAppointDocLabel;//Per doctor
+>>>>>>> parent of 54654e0... Updating Authority
 	
 	private JPanel contentPanel;
 	private JPanel statPanel;
-
 	
-	private int depVal;
-	private int patientVal;
-	private int docVal;
-	private int nurVal;
+	
 	
 	
 	
@@ -74,7 +71,11 @@ public class AuthorityView extends JFrame{
 	{
 		//sets frame containers attributes
 		setTitle(title);
+<<<<<<< HEAD
 		setSize(470,360);
+=======
+		setSize(600,500);
+>>>>>>> parent of 54654e0... Updating Authority
 		setDefaultCloseOperation(DISPOSE_ON_CLOSE);
 		setLocationRelativeTo(null);
 		
@@ -104,31 +105,18 @@ public class AuthorityView extends JFrame{
 
 
 
-		departBaseLabel = new JLabel("# of Departments: ");
+		departBaseLabel = new JLabel();
 			departBaseLabel.setFont(new Font("Tahoma", Font.PLAIN, 13));
-		departValLabel = new JLabel(""+depVal);
-			departValLabel.setFont(new Font("Tahoma", Font.PLAIN, 13));
-
-
-
-		patientTotalLabel = new JLabel("# of Registered Patients: ");
+		
+		patientTotalLabel = new JLabel();
 			patientTotalLabel.setFont(new Font("Tahoma", Font.PLAIN, 13));
-			
-		patientValueLabel = new JLabel(""+patientVal);
-			patientValueLabel.setFont(new Font("Tahoma", Font.PLAIN, 13));
-
-
 
 		regDoctorTotalLabel = new JLabel("# of Registered Doctors: ");
 			regDoctorTotalLabel.setFont(new Font("Tahoma", Font.PLAIN, 13));
-		regDoctorValLabel = new JLabel(""+docVal);
-			regDoctorValLabel.setFont(new Font("Tahoma", Font.PLAIN, 13));
-
-
-
 
 		regNurseTotalLabel = new JLabel("# of Registered Nurses: ");
 			regNurseTotalLabel.setFont(new Font("Tahoma", Font.PLAIN, 13));
+<<<<<<< HEAD
 
 		regNurseValLabel = new JLabel(""+nurVal);
 		regNurseValLabel.setFont(new Font("Tahoma", Font.PLAIN, 13));
@@ -145,22 +133,25 @@ public class AuthorityView extends JFrame{
 			totalAppointLabel.setFont(new Font("Tahoma", Font.PLAIN, 13));
 
 		totalAppointDocLabel = new JLabel("# of Doctors today: ");
+=======
+				
+		totalAppointDocLabel = new JLabel("# of Scheduled appointments for Specific Doctor: TBA");
+>>>>>>> parent of 54654e0... Updating Authority
 			totalAppointDocLabel.setFont(new Font("Tahoma", Font.PLAIN, 13));
-		totalAppointNurLabel = new JLabel("# of Nurses today: ");
-			totalAppointNurLabel.setFont(new Font("Tahoma", Font.PLAIN, 13));
 
-			
-
-		
-
-
+		totalAppointDepLabel = new JLabel("# of Scheduled appointments for Department: 0");
+			totalAppointDepLabel.setFont(new Font("Tahoma", Font.PLAIN, 13));
 
 		statPanel= new JPanel(new MigLayout("wrap 2", "[] 16 []") );
 			statPanel.setBorder(BorderFactory.createTitledBorder("Statistics here"));
 
 
+		//Scroll bar
+		departmentDropDown = new JComboBox<String>();
+		departmentDropDown.insertItemAt("", 0);
 
 		
+<<<<<<< HEAD
 		statPanel.add(departBaseLabel);
 		statPanel.add(departValLabel, "wrap");
 	
@@ -186,10 +177,19 @@ public class AuthorityView extends JFrame{
 
 		
 		
+=======
+		statPanel.add(departBaseLabel, "wrap");
+		statPanel.add(patientTotalLabel, "wrap");
+		statPanel.add(regDoctorTotalLabel, "wrap");
+		statPanel.add(regNurseTotalLabel, "wrap");
+		statPanel.add(totalAppointDocLabel, "wrap");
+		statPanel.add(totalAppointDepLabel, "wrap");
+>>>>>>> parent of 54654e0... Updating Authority
 
 		contentPanel.add(welcomeLabel, "center, wrap");
 		contentPanel.add(btnReturn, "center, wrap");
-		contentPanel.add(statPanel, BorderLayout.CENTER);
+		contentPanel.add(statPanel, "center, wrap");
+		//contentPanel.add(departmentDropDown); //Temporarily locked
 
 		setVisible(true);
 	}
@@ -202,46 +202,39 @@ public class AuthorityView extends JFrame{
 		this.welcomeLabel = welcomeLabel;
 	}
 	
-	public JLabel getDepLabel(){
-		return departValLabel; 
-		
+	public JLabel getDepNumLabel(){
+		return departBaseLabel; 
 	}
 
-	public void setDepLabel(JLabel label){
-		departValLabel = label;
+	public void setDepNumLabel(JLabel departBaseLabel){
+		this.departBaseLabel = departBaseLabel;
 	}
 
-	public JLabel getPatientLabel(){
-		return patientValueLabel; 
+	public JLabel getPatientTotalLabel(){
+		return patientTotalLabel; 
 	}
 
-	public void setPatientLabel(JLabel label){
-		patientValueLabel = label;
+	public void setPatientTotalLabel(JLabel label){
+		patientTotalLabel = label;
 	}
 
-	public JLabel getDoctorLabel(){
-		return regDoctorValLabel; 
+	public JLabel getDoctorTotalLabel(){
+		return regDoctorTotalLabel; 
 	}
 
-	public void setDoctorLabel(JLabel label){
-		regDoctorValLabel = label;
+	public void setDoctorTotalLabel(JLabel label){
+		regDoctorTotalLabel = label;
 	}
 
-	public JLabel getNurseLabel(){
-		return regNurseValLabel; 
+	public JLabel getNurseTotalLabel(){
+		return regNurseTotalLabel; 
 	}
 
-	public void setNurseLabel(JLabel label){
-		regNurseValLabel = label;
+	public void setNurseTotalLabel(JLabel label){
+		regNurseTotalLabel = label;
 	}
 
-	public JLabel getTotalAppointLabel(){
-		return totalAppointLabel;
-	}
 
-	public void setTotalAppointLabel(JLabel label){
-		totalAppointLabel = label;
-	}
 	
 	public JLabel getActualAppDayLabel(){
 		return actualAppDayLabel;